Fix false sharing issues
In this commit we fix the false sharing issue by using an
alignment factor of 64 bytes so that the atomic values
that are used by the individual threads are not falsely
shared. We also use a function local value to perform
the accumulation locally and only then update the atomic
value in order to reduce atomic instructions executed.